requirements.txt
flask
pyyaml
requests
logging-gelf
pymssql
pymssql
